OM-39327: XL: Remove external-api-stub module from XL.
Intent
======
Move the external-api-stubs module from XL into opsmgr, so that it lives together with the controllers it generates stubs for.

This is the XL-side of that change.

Background
==========
We get fairly regular issues in this module when there are problematic checking with com.vmturbo.api that don't play well with the swagger-maven-plugin we use to generat a swagger.json file for our controllers/DTOs. These problems should be detected and remedied at the time of the com.vmturbo.api checkin.

Testing Done:
- Build XL after building the opsmgr side of this change locally.

Reviewed By: jeff.bischoff
https://rbcommons.com/s/VMTurbo/r/26232/