OM-39352: Call CPU Capacity API from API component to fetch the list of known cpu_models RB-26310: Karthik, Patrick
background
The UI displays a list of available CPU Models in the "create host template" panel. This list of cpu model names comes from the 'cpucapacity' library. In a previous change, OM 39143, we called this library from the API component to fetch the list of names to return in response to the REST request.

In OM 39351 this library, 'cpucapacity', was moved to Plan Orchestrator as a better design, and gRPC calls added to retrieve the CPU Library information.

changes
In this review we change the API component to request the list of CPU Models from the Plan Orchestrator, rather than getting the list directly from the 'cpucapacity' library.