OM-39393 : Set IgnoreConstraint in scenario plan changes and deactivate commodities based on the ignoreconstraint setting. RB: 26453
IgnoreConstraint is sent as an argument in the REST api during plan execution
instead of as part of the scenario change. The ignoreConstraint argument was not
being ignored in the XL api component which lead to wrong plan results.

Since no customer is using this yet, making breaking changes to the PlanDTO
protobuf is fine.