OM-39531: XL: UCS-Director category is shown with all caps - 'ORCHESTRATOR' RB-26499: Roman, Christine
background
Each probe_config.xml declares the category for the probe. This is completely up to the probe writer, and so the writer for UCS-D used "ORCHSTRATOR". There is a mapping step in the classic api service for targets/probes which maps this value back to a user-friendly string.

solution
We use the same approach and map a small number of "ugly" probe category strings to user-friendly ones. This map contains the categories we know of, and so may need to be re-visited in the future - or we can ensure that new probe writers use user-friendly category names to begin with.

Note that there is at least one existing probe category ("Guest OS Processes") that would not allow us to use a simple "capitalize-first-letter" translation.