OM-39059: XL: Support search filters by "state" in the TP.
Intent
======
To make filters by the "state" property work in the Topology Processor. Currently attempting to filter entities by state throws an IllegalArgumentException.

Background
==========
Policies on dynamic groups that have a "state" filter were not being applied because the Topology Processor failed to resolve the "state" filter in it's group resolution code. Adding support for this is fairly straightforward (see: **TopologyFilterFactory**).

The tricky part of this change is that the search filter's value is in UI/API terms - "ACTIVE", "IDLE", and so on. We can't convert it to XL terms (EntityState enum) in the API component, because the repository stores the state in UI/API terms, and searches that go to the repository won't work with EntityState. Thus, the bulk of this change is moving the state mapping code into a shared location in com.vmturbo.components.common. From there it's used by Repo, TP, and API.

Note: There are states that we don't currently handle in XL. My impression is that these are not related to the discovered "power" state, but depend more on runtime information about the entity and/or actions associated with the entity. We'll need to add support for that at a later point.

Testing
=======
- Full build.
- Started up components. Created dynamic group, filtered by name and state. Applied placement policy to group. Verified action was generated.

Reviewed By: emanuele.maccherani, karthikt
https://rbcommons.com/s/VMTurbo/r/26397/