OM-39568: XL Market/Cost: Adding StorageCostBundle to MarketPriceTable.
Intent
======
Allow entity -> trader conversion to include pricing data about storage tiers in the topology.

Background
==========
The overall logic is the same as for compute - we loop over the business accounts for a particular (region, tier) combination and create individual price items for each account, considering the discount for that account.

The one tweak is that storage tiers usually have staggered pricing - meaning that the price for each tier may contain several prices, with "end_range" specified.

Note 1 - Difference from Classic
================================
In classic, "isUnitPrice" and "isAccumulative" price is explicitly set based on the type of storage. In XL, we can infer it from the units of the price, and the presence of ranges. We don't need to explicitly consider the storage type - the cost probe is responsible for setting the price list propertly for each tier.

Note 2 - Divergence from Classic
================================
For the Azure price buckets (STORAGE_PERGB_COST_FOR_SIZE) we will use non-unit price, because based on my understanding each GB range has a flat cost. For example:
- $5 for the first 10GB
- $4 for the next 10GB
This means that regardless of how much you consume in the (0, 10GB) range, the price is $5.

Tests
=====
Unit testing, with additional unit tests. Can't test end-to-end yet, because not all the pieces are in place.

Reviewed By: wei
https://rbcommons.com/s/VMTurbo/r/26418/