OM-39250: Add the scaling factor to the CPU commodity during topology broadcast RB-26486: Ron, Pengcheng
background
We have created a library, com.turbonomic.cpucapacity, which given a CPU Model name returns a scaling factor relative to a known CPU baseline using SPEC benchmarks. In a prior task, OM 39145, we copy a CPU Model name from a template into a templated PM during topology-editing, as part of plan execution.

problem
During topology broadcast we add copies of a template, if requested, to the plan topology to be broadcast. Each PM template with a CPU Model specified should have the scalingFactor for that CPU Model applied to:
1) the CPU capacity (sold)
2) the CPU_Provisioned
3) the VCPU sold by any VM hosted by this PM

solution
- add 'scaling_factor' to TopologyDTO.proto, CommoditySoldDTO; also to CommoditySoldRepoDTO
- add the 'scalingFactor' field processing to the Repository code to store/retrieve ServiceEntityRepoDTO's
- add CpuScalingFactorPostStitchingOpereation to implement the application of the scalingFactor during topology processing; also add this new operation to the list in PostStitchingOperationLibrary so it will be applied
- add an interface CpuCapacityStore to be called from the stitching operation - to implement fetching scalingFactor for a given cpuModel
- implement RemoteCpuCapacityStore in TopologyProcessor - to be called from the stitching library and fetches the scalingFactor from Plan Orchestrator where the cpu spec library is hosted; this Store also includes local caching to reduce chattiness; the cache time-to-live is configured from a new external setting (added to factoryInstalledComponents.yml)