OM-37329: XL: Calculate and broadcast projected entity costs.
Intent
======
We should be able to retrieve per-entity cost projections (with cost breakdowns by category) for the projected topology.

Background
==========
The market produces a projected topology as one of the analysis outputs. In order to see the improvements in a cloud environment we want to see the per-entity costs in the projected topology.

In this change we are using the cost calculation library to get the projected costs. The market should also have costs on the Traders, and another option is to use those costs for the projections. The upside of that approach is that it's more efficient, and the costs are more likely to be consistent with the savings in the ActionTOs generated by the market. The downside is that there are no cost breakdowns by category, and the costs are more likely to be inconsistent with the "before" costs, since they are not being calculated by the same mechanism. In any case, since the cost calculation is internal to the market component, it should be easy to switch the method without affecting the rest of the system.

This change adds the following flow:
1) **Analysis** Accept **TopologyCloudCostCalculator**, **TopologyEntityCloudTopologyFactory**, and **MarketPriceTableFactory**. Constructing a cloud topology and market price table. Calculating projected entity costs after analysis completes. We save the original cloud topology and the projected entity costs because we will need it to calculate "before" costs for action savings.
2) **MarketRunner** Get the projected entity costs from the completed analysis.
3) **MarketRunner/MarketNotificationSender** Broadcast the projected per-entity costs via Kafka.
4) **CostComponentProjectedEntityCostListener** Receive the projected costs and store them in **ProjectedEntityCostStore**. For now we store them in a simple map, until the expected usage patterns become more clear.

The majority of the changed files are around adding support for a **projected-entity-costs** topic to Kafka.

Note: We only calculate/broadcast projected per-entity costs for realtime topologies for now, until there is more clarity on what kind of cost information we need to keep for market.

Minor change:
CloudTopology::getConnectedService will return the service if given a service ID.

Tests
=====
- Ran full build.
- Started up components and verified that a fake entity cost message gets delivered from the market component to the cost component.

Reviewed By: thiru
https://rbcommons.com/s/VMTurbo/r/26252/