OM-38107: XL Cost: Store price tables in MySQL
Intent
======
To persist price tables in MySQL instead of in memory.

Background
==========
We want to be able to store the price tables discovered by the cost probes in a database instead of in memory. This becomes particularly important when/if we desynchronize the upload schedule of the price tables from the broadcast schedule.

Adding a new PriceTableStore implementation (**SQLPriceTableStore**) that stores the price tables (by probe type) in a new "price_table" table. The tables are stored as blobs, because we don't need to run queries on the contents of the table.

We also store ReservedInstancePriceTable protobufs in the PriceTableStore, for use by the BuyRI algorithm.

Note
====
I'm changing the upload price table call to include reserved instance prices, and to upload a map of price tables by probe type. This change is inspired by Patrick's changes, just without the checksum.

Testing
=======
Unit testing, with additional tests.

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/26289/