OM-38084: XL Group: Allow groups with the same display name if entity type is different.
Intent
======
Two groups that have the same display name but different types of entities should be allowed in the system. This is especially important for clusters.

Background
==========
In the travelport environment we have clusters and storage clusters with the same name, and we are failing to discover one of them. We already use (name, entity type) to identify the group in TargetCollectionUpdate, but the underlying GroupStore used to throw exceptions when the name was the same.

This change:
1) Removes the UNIQUE constraint on the name.
2) Adds entity_type to the grouping table.
3) Adds migration to the group component (see everything in the group/migration folder). All entity_type rows start with value -1, which is not valid. The migration sets the entity_type row to the entity_type in the group blob.

The main changes are in the new migration class. We do this so that the migration can target only the rows that need to be changed, instead of relying on a possibly changing method in the GroupStore. We delete rows where the group_data is invalid, because we can't get a valid entity type out of it - and it has no user impact, since we wouldn't be able to return those groups to the user anyway.

Tests
=====
- Unit tests.
- Tested migration in the group component:
    - Built 7.6.0 (without this change)
    - Started group component, manually added 3 groups (VM, Host, and Invalid) via swagger. To create an invalid group, manually corrupted the group_data column.
    - Verified the entity tables are pre-migration (no entity type column)
    - Built develop.
    - Restarted group component. Verified the VM and Host groups have their entity types set, Invalid group is deleted.
    - Looked at the logs - batch was of size 1 for testing.

Reviewed By: karthikt
https://rbcommons.com/s/VMTurbo/r/26529/