OM-39936: XL Cost: More forgiving/correct logic for letting entities into the cloud topology.
Intent
======
Reduce the likelihood of an empty/incomplete cloud topology.

Background
==========
There are error cases - as well as legitimate cases - where we may not be able to get the target/probe information from the TP, or where we may not be able to get the discovering targets of an entity (e.g. for entities created for a plan). The main changes:
1) If an entity does not have discovering targets, include it in the topology.
2) If we fail to retrieve probe-target mappings, include all entities in the topology.

It's should be relatively harmless to let non-cloud entities into the cloud topology - at least compared to excluding cloud entities from it. If entities further down the chain (e.g. the stats writer) need to perform additional checks, then we will add logic to do that later on.

Testing
=======
Unit testing.

Reviewed By: dblinn
https://rbcommons.com/s/VMTurbo/r/26654/