OM-39146: Apply the cpu model scaling in TopologyConverter RB-26519: Nitya
background
Using the 'cpu capacity' library we calculate a scaling factor for the CPU and CPU CAPACITY commodities sold for PM templates based on the cpu model for that template (if indicated by the user). This 'scalingFactor' should be used to multiply-up the corresponding commodities on the way "in" to the market analysis, and then divided-down again on the way back out of the market analysis.

solution
- We added a check for a scalingFactor when converting commodities-sold from the XL model object CommoditySoldDTO to the market model object CommoditySoldTO. This scaling factor is also used to convert CPU CommodityBought commodity for VMs with 'scalingFactor' set.
- Conversly, we added a check for a scaling factor when converting back from CommoditySoldTO (market) to CommoditySoldDTO (XL) and CommodityBoughtTO (market) to CommodityBoughtDTO (XL).