OM-38084: XL Group: Look at "ENTITY_TYPE" instead of "TYPE" when checking for duplicate groups.
Intent
======
Allow uploads of groups with the same name but different types.

Background
==========
Even though I added most of the code to support different-type groups, the query checking for duplicates was looking at the "TYPE" column - which indicates the type of the group - instead of the "ENTITY_TYPE" column.

In the unit tests, I unfortunately set the "ENTITY_TYPE" of the group to 1 - which happens to be the value associated with the "Group.Type.GROUP" enum. So the test passed even though it was looking at the same column.

The fix: look at the ENTITY_TYPE column. I also changed the entity type in the test, and verified that the unit test fails before this fix and succeeds after.

Testing
=======
- Unit test now fails before the change.
- Tested with DC2 - I was able to see the DC2 cluster and storage cluster.

Reviewed By: emanuele.maccherani
https://rbcommons.com/s/VMTurbo/r/26671/