OM-38104: XL Cost: Calculate Storage Cost for volumes and VMs.
Intent
======
Calculate storage cost for volumes, and VMs attached to those volumes.

Background
==========
In XL, we want to show storage costs on both volumes and VMs.
VMs are connected to volumes (via a "NORMAL" connection). Volumes are connected to storage tiers (via a "NORMAL" connection), and the capacity of the volume is set in the volume's type-specific info.

To calculate the storage cost for a volume, we use the pricing information of the connected-to storage tier and the IOPS/GB capacity of the volume. See: **CloudCostCalculator::calculateVirtualVolumeCost**.

To calculate the storage cost for a VM, we get the connected volumes, and store them in the VMs cost journal. When the cost calculation for all entities is complete and we want to get the costs for the VM's cost journal, we factor the volume's cost journal into the VM's cost journal. See: **CostJournal::calculateCosts**.

Testing
=======
- Additional + Existing unit tests.
- Some end-to-end testing: saw some storage costs in the entity_costs table.

Reviewed By: emanuele.maccherani
https://rbcommons.com/s/VMTurbo/r/26645/