OM-39941: XL TP: When replacing entities, include replaced entity in new entity's display name.
Intent
======
Entities replaced with template in a plan should have names that indicate the entity they replace.

Background
==========
The actual issue was that replacing multiple entities with the same template was causing only one entity to appear in the "after" section of the optimized improvements widget. This happened because the widget seems to arrange entities by display name at some point. All replacements had the same display name, so only one entity appeared.

The solution is to generate semi-unique display names - in particular, the name of the replaced entity should be in the display name of the replacing entity. This makes sense for action visualization as well - it becomes easier to distinguish different entities of the same template.

We should probably also change the widget to use id instead of display name for de-dupes (since display name is not guaranteed to be unique) but that can happen later.

Testing
=======
- Build TP.
- Ran hardware refresh plan, replacing 4 hosts. All 4 replacements appear in optimized improvements widget, with appropriate names.

Reviewed By: mark.laff, pengcheng.wang
https://rbcommons.com/s/VMTurbo/r/26875/