OM-40361: XL Cost/Market: Refactoring TopologyCostCalculator to make it's CloudCostData accessible.
Intent
======
The market component should only retrieve CloudCostData once, and re-use it in all the places it needs.

Background
==========
We're working on adding addition features (e.g. RI projections, CBTP creation) that need access to data in the cost component. We are already retrieving this data for cost calculation. The retrieval of data is a pretty expensive operation, so we shouldn't be doing it multiple times.

This change modifies the flow in the following way:
- TopologyCostCalculator is no longer a singleton created in Spring. It's created by TopologyCostCalculatorFactory (new), and fetches the current CloudCostData when it's created.
- Analysis accepts TopologyCostCalculatorFactory as an argument (instead of TopologyCostCalculator). Then:
    - Create a new instance of TopologyCostCalculator to use for the analysis.
    - Pass the TopologyCostCalculator's CloudCostData to the MarketPriceTableFactory (eliminating an extra call to get CloudCostData)

From now on, if we need CloudCostData anywhere in Analysis (e.g. during conversion, or action interpretation) we can use TopologyCostCalculator::getCostData to get the same version of the cost data that we'll use to calculate projected costs and action savings. This is a big win for both consistency and performance.

Testing
======
- Ran tests in the market and cost component.

Reviewed By: thiru
https://rbcommons.com/s/VMTurbo/r/26946