OM-40390: Add necessary type-specific info fields to TopologyDTO.proto RB-26996: Roman
Background:
A TopologyEntityDTO protobuf has a TypeSpecificInfo field (called 'type') which is a oneof containing a structure specific to a single ServiceEntity type.

Change:
In this commit we are adding new TypeSpecificInfo sub-messages for ApplicationInfo and StorageInfo to capture the entity-specific fields for each of those SE types. We are also adding 'vendor' and 'model' to the pre-existing PhysicalMachineInfo.

This change just adds new messages or fields. Code to make use of these fields will follow.