OM-39917: XL Action: Update the internal ActionDTO.Action of a re-recommended Action to reflect latest properties.
Intent
======
Re-recommended actions should reflect the most recent properties recommended by the market.

Background
==========
When an action with the same ActionInfo is re-recommended by the market, we don't create a new Action object for it in the Action Orchestrator so that we can keep the action addressable by the same ID. However, there are properties outside the ActionInfo that can change between recommendations. If we use the same Action object we need to update those properties, because the difference can be significant. For example, the explanation or importance may change, or the action may become executable where before it wasn't (or vice versa).

Implementation
==============
- Action.recommendation (ActionDTO.Action) and Action.actionTranslation (ActionTranslation) are no longer final.
- **Action::updateRecommendation** overwrites both the recommendation and the action translation. It may be safe to keep the action translation for the existing translations, but conceptually any part of the recommendation can affect the translation.
- **LiveActionStore** calls **updateRecommendation** on existing re-recommended actions.

Reviewed By: emanuele
https://rbcommons.com/s/VMTurbo/r/26857/