OM-40353: XL Repo/API: Allow restricting supply chain query by Environment Type.
Intent
======
We should be able to show different supply chains for "cloud", "on-prem", and "hybrid" environments.

Background
==========
We are adding the environment type to entity information in the repository in a separate change. However, the repository's SupplyChainService needs to be able to use the environment type to return subsets of the "total" supply chain. This is most relevant in the global overview - i.e. with the "Market" scope. In most other cases we will implicitly be restricted to either cloud or on-prem.

Main changes:
- **SupplyChain.proto** - the request for the supply chain now includes an optional environment_type.
- **ArangoDBQueries, ReactiveArangoDBExecutor, and ArangoDBExecutor** - changes to support the environment type for global and single-source supply chains. Switched to using ST templates instead of the apache Substitutor so that we could use conditionals in the template.
- **SupplyChainFetcherFactory** - send the environment type as part of the request, if it came in via the API.

Testing
=======
Tested an environment with an AWS and VC target.
- Saw different supply chains in "ON PREM", "HYBRID", and "CLOUD", with appropriate entity counts.
- Temporary groups were created properly when clicking on circles in the supply chain.

Reviewed By: pengcheng
https://rbcommons.com/s/VMTurbo/r/26922/