OM-40511: Add REST endpoint to set the defaults for a Component Type RB-27076: Patrick
Background
The default configuration properties for all XL components have, before this, all been set in a single file built into Clustermgr: 'factoryInstalledComponents.yml'. This means that adding a component or changing default configuration properties for a component requires rebuilding / redeploying /re-running Clustermgr - which is unacceptable.

Fix
This review is the first step towards moving the default configuration properties from 'factoryInstalledComponents.yml' to each Component. In this review we simply add an entrypoint to the Clustermgr REST API and the ClusterMgr API to call it.  The base method to implement this call (clusterMgrService.putDefaultPropertiesForComponentType) was already available.

In a future review we will change component initialization to call this entrypoint to establish defaults.