OM-40392: Fill in TypeSpecificInfo fields from the corresponding SDK EntityDTO RB-27008: Roman, Jeff, Christine
Background
We have defined a number of TypeSpecificInfo messages within TopologyEntityDTO. The TypeSpecificInfo for each entity type holds fields unique to that entity type. This information is used to (a) display in the UI, and (b) be used for grouping / searching.

Implementation
We add an interface TypeSpecificInfoBuilder that defines a map TYPE_SPECIFIC_INFO_BUILDER from EntityDTO.EntityCase to a Builder class for that entity type. Each builder constructs an instance of the relevant TypeSpecificInfo and populates the fields from the given EntityDTO.

Note that not all TypeSpecificInfo structures are defined, and not all type-specific info fields have been added / populated. This framework will be expanded in future updates as needed.