OM-40391: Store and Retrieve TypeSpecificInfo to/from the Arangodb repository RB-27010: Pengcheng, Ioannis
Background
The "ServiceEntityRepoDTO" class is the bean persisted-to and retrieved-from Arangodb. This structure already had fields for type-specific info for VirtualMachine and ComputeTier.

Change
This commit adds type-specific info for ApplicationInfo, DatabaseInfo, PhysicalMachine, StorageInfoRepoDTO, VirtualVolume. We also split the converters for "ServiceEntityDTO -> RepoDTO" and "RepoDTO -> ServiceEntityDTO" into separate classes, and change them into table-driven converters based on the entity type. The actual converstion methods to/from RepoDTO are in each individual xxxInfoRepoDTO bean, e.g. VirtualMachineInfoRepoDTO.

Also: rename the search service RPC class "SearchRpcService" (was just "SearchService") and include the class in the swagger-maven-plugin config so the Search RPC is included in the generated swagger ui.