OM-40389, RB-27118: Implement API Component translation from TopologyEntityDTO to Aspects
background
For each entity type there are number of 'aspects' defined containing one or another portion the "external data" representing an entity. Some of these aspects are shared by different entity types, as there may be an overlap between the external data items that pertain to each of the entity types.

Solution
We extend the prior structure of ASPECT_MAPPERS to include all the entity types available in class. We create an implementation of IAspectMapper for each of the aspect types, although most of those are stubs for now.

There is a following JIRA, OM 40635, to populate all of the fields in all of the aspect matchers.