OM-40708: Consul k8s deals with reboot RB-27247
background
consul keeps track of a list of peer servers in the "raft", and when one consul image comes back up it tries to contact the others to "get back in the game". If there is only one consul server, and its IP address changes from invocation to invocation, the new one keeps trying to contact the old one to "vote for a leader"... and the old one will never answer.

fix
During startup consul looks for a file "/consul/data/raft/peers.json". If this file exists, then consul uses this to locat the peers instead of the prior saved list. If we simply write the current IP to this file during consul container startup then the vote can proceed correctly. After consul starts up it automatically removes the 'peers.json'.

Note that there is a comment on the web about the 1.4 version of Consul handling the case where all the peers in the raft get new IP addresses, but we tried updating to consul version 1.3 and our problem still persisted.  That's why we need change to create a 'peers.json' file.