OM-37768: fix XL post install target addition sticks at discovering (rb27232)
Problem:

During onboard mode (in UX term), after successfully adding the first target and waiting for several minutes, the 'Discovering' splash is still displayed even though the system is ready and the supply chain should render. A manual browser refresh results in a broken image. The user then have to click on the "On" button or relogon to  the UI to return the appropriate supply chain.
"Discovering" splash
https://vmturbo.atlassian.net/secure/attachment/89945/89945_Screen+Shot+2018-08-03+at+10.49.06+AM.png
"broken" image
https://vmturbo.atlassian.net/secure/attachment/89944/89944_Screen+Shot+2018-08-03+at+10.48.54+AM.png

Investigation:

In 6.x, a new system notification feature was added, the main function is:
During onboard mode, when the discovery has finished in the backend, a message through the websocket channel is sent to the UI, that will switch to normal mode and trigger the redirect to main page.

But this feature is not implemented in XL.

Design:

1. Add new system notification API to XL.
2. Send a websocket notification to UI when backend is "ready".

It's tricky to determine #2. In XL "ready" means,
a. target discovered
b. "topology broadcast" triggered by scheduler (10 minutes by default)
c. market received and analysed the topology, broadcasted the results
d. other components (e.g. action and history) processed the results.

Also it's preferable not to send the system notification to UI after onboarding.

Implementation options:

Precondition: after adding the first target
1. wait for target discovered
2a. wait for "topology broadcast" (could be up to 10 mins by default).
Or
2b. trigger "topology broadcast"
3a. wait for all the components to finish processing results
or
3b. wait for one component
4. send system notification to UI
Postcondition: stop sending subsequence notifications.

Chosen implementation option:

1. wait for target discovered
2b. trigger "topology broadcast"  - better usability, low system impact (only one target)
3b. wait for two components (repository and history): repository provide supply chain and history component provides most data for UI widgets.
4. send system notification to UI