OM-40721: XL API: Handling "All On-Prem Hosts" group in stats request, and not erroring on environment type stats filters.
Intent
======
The "Risk Index", "Headroom" and "Optimized Improvements" widgets on the main on-prem window shouldn't get API errors.

Background
==========
When you click on the "On Prem" tab, the default widgets use the "ALL_ON_PREM" scope - which has a hard-coded UUID on the UI side, to match a hard-coded UUID in classic's pre-created groups. We don't have pre-created groups in XL, and even if we did the UUID would be different.

The solution is to create a temporary group - basically the same thing we do when you click on the supply chain donut - for all PM hosts in the ON_PREM environment (which is actually equivalent to just all PM hosts) - and to replace the "magic" UUID with the OID of that group. We also need to invalidate the temporary group when it's members are no longer valid, or when it expires in the group component.

Implementation
==============
**MagicScopeGateway** - responsible for taking UUIDs coming in from the market, and replacing "Magic" scopes with OIDs that can be properly interpreted by the rest of the system. Right now this only works for ON_PREM_HOSTS, although we may want to do something similar for other magic groups.
    - Each time we access the temporary group for the ON_PREM_HOSTS we check if it still exists. If it doesn't, we re-create it.
    - The gateway also listens for new source topologies available in the repository, and manually invalidates the mapping when a new source topology is available. This way we're always working with "up-to-date" data.

Other/Supporting Changes
========================
- Adding EnvironmentType to TempGroupInfo.
- Not applying global scope optimization to TempGroups with EnvironmentType == CLOUD. We still apply it when type == ON_PREM even though it's not quite correct, because the performance impact is significant.
- Handling EnvironmentType filter on stats in the history component. For now we just ignore it, to avoid errors.

Testing
======
- Hands-on testing with XL and stress-probe targets.
    - Navigated to "On Prem" view, verified there are no more errors and headroom chart renders.
    - "Risk Index" chart only has one data point (projected) because we don't currently aggregate price index (so the market stats tables don't have an entry for it). This should be handled separately.
    - "Optimized Improvements" widget doesn't have data because of a UI bug - it's requesting the wrong stuff.
    - Clicked around on the supply chain. As expected.
- Unit tests.

Reviewed By: garyzeng
https://rbcommons.com/s/VMTurbo/r/27324/