OM-40961: 'turboctl.py logs' no longer works, RB-27409 Background the 'turboctl.py logs' subcommand builds on the 'docker logs' command, fetching the log for the 'rsyslog' container and then grepping for the desired container names, if any.
Problem
the most recent version of docker-compose adds the container ID to the container name as listed by 'docker ps'. This make the 'rsyslog' container's name look like: "build_rsyslog_1_afb7345c870c". In order to feed the container name to the 'docker logs' command we were using the name "build_rsyslog_1", which is no longer the container name.

Solution
We ask for all the docker container names, looking for the 'rsyslog" pattern. The first token on the line is the unique ID for the rsyslog container. We pass that to the 'docker logs' command instead of the character-based container name. In this way the 'docker logs' command will always operate on the correct container even if the name changes format.