OM-41132: Repository: NPE during broadcast (related to setVendor); RB-27536
Background:
We store topologies of ServiceEntities in the Repository as ServiceEntityRepoDTO's. Several types of ServiceEntity have a set of attributes unique to that Service Entity Type, e.g. PhysicalMachineInfoRepoDTO. The type-specific subfield (e.g. PhysicalMachineInfoRepoDTO) must be (a) populated from a broadcast topology, and (b) turned back into a TypeSpecificInfo protobuf on request for an entity or topology.

Problem:
If a field of the TypeSpecificInfo is not specified, then the corresponding field of the RepoDTO will be null and stored in the database (Arango). The code to turn that RepoDTO back into a protobuf needs to check for nulls.

Solution:
Add null checks to the conversions back to protobuf for each Entity Type specific RepoDTO.
Also add a null check to the ApplicationInfoRepoDTO conversion - no bug detected yet, but we should add the check to be future-proof.