OM-33192: Hyper-V probe may sometimes generate clashing local IDs for VMs and Applications; RB-27502
Background
Object Identity is keyed on attributes specified in the configuration for each probe. The file is called 'identity-metadata.yml'

Problem
In the hyperv the 'identity-metadata' specifies "id", and "entityPropertiesList(name,clusterInvariantUUID)/value*" as the primary identifying characteristics. Unfortunately, there seem to be cases where these values are the same between a VM and an APPLICATION.

The Fix applied in OM 33546 was to add "entityType" to the identity-metadata.yml for HyperV (and VMM which shares base code). This supplied the immediate differentiator and fixed the problem. However, that approach puts the onus on the probe writer to add the "entityType", and as the problem occurs infrequently it may occur in the field for some other probe.

Solution
Instead of having to specify "entityType" for every 'identity-metadata.yml' file we automatically add "entityType" as a primary identifying characteristic. In other words, if two SE's have different EntityType they will never be "the same" for OID generation.