OM-40062 Populate Projected RI Coverage and RI Utilization widgets. RB:27671
Display the Projected RI Coverage and Projected RI Utilization widgets
in the Cloud UI.

Added a kafka listener to the Cost component for subscribing to the
Projected RI Coverage messages from market. On recieving the RI coverage
messages, they are stored in-memory in the cost component. When the RI
Coverage and Utilization stats are requested from the UI, the api calls
the Cost Component RPC to fetch the projected RI stats from the
Projected RI Coverage Store.

Also fixed some of the cost component unit tests which were not cleaning
up the DB at the end of the unit test execution.