OM-40923: XL History: Aggregating price index by entity type in Market Stats tables.
Intent
======
Aggregate market-scope Price Index in the market_stats tables to allow fast lookups of global price index. Fix Risk Index widget for the global "on-prem" view.

Description
===========
We currently save price index to the individual tables (e.g. pm_stats, vm_stats), but we don't save the aggregate price index to the market_stats tables. That means price index queries for the global scope don't return any data - and if we want to retrieve market-wide price index we need to query each individual entity.

The fix is to aggregate the price index (in the same way that we aggregate all the other stats), and insert a price index row for each entity type into the market_stats_latest table when processing the projected topology in the history component.

Major changes:
- **TopologyDTO.proto** - The list of skipped entities in the projected topology now includes the entity type. The entity type is often required when processing OIDs.
- **PriceIndexWriter** now aggregates the price index and writes those extra rows to the market_stats_latest table. Skipped entities get a default value of 1.0.

Refactoring
===========
Now that we have the price index (and the skipped entities' types) in the projected topology, we don't need the TopologySnapshotRegistry and TopologyOrganizer classes. I removed them, and made the necessary refactoring - most notably replacing uses of TopologyOrganizer with TopologyInfo, which contains the same fields.

Testing
=======
- Full build, with tests.
- Added unit tests for PriceIndexWriter (and found a minor off-by-one bug :) )
- Started components. Verified that the Risk Index widget works for the global "on-prem" view.

Reviewed By: mark.laff
https://rbcommons.com/s/VMTurbo/r/27852/