OM-41516: XL API: Fetching GroupApiDTO when groups are used as scope for WidgetApiDTOs loaded from DB.
Intent
======
Custom Pending Action widget scoped to group should show actions (if entities in the group have any). Generally, saving and loading a dashboard widget (i.e. a WidgetApiDTO) should not lose important GroupApiDTO data - for example, the type of entities in a group.

Background
==========
A custom Pending Actions widget scoped to a group would not show any actions. This was because the "groupType" of the widget's scope was not set in the API response. It's not set because the WidgetApiDTO::scope field is a "BaseApiDTO" - not a generic type - and the Jackson serializer we use to save the DTOs to the database drops all fields that are not part of "BaseApiDTO."

To fix this, I added a post-processing method when deserializing the "scope" field of WidgetApiDTO. If the scope's class is some group (Group, Cluster, Storage Cluster) then we retrieve the group from the group component, map it to a GroupApiDTO, and set the "scope" field to that GroupApiDTO.

Note: I also made WidgetsetMapper non-static, and made the relevant changes to WidgetSetsService.

Testing
=======
- Tested in the UI - verified the custom Pending Actions widget scoped to a group now shows actions.
- Ran existing tests.
- New unit tests for WidgetsetMapper.

Reviewed By: jeff.bischoff, mark.laff
https://rbcommons.com/s/VMTurbo/r/27895/